Added retries to the job completion handler.
