+++ genie-web/src/main/java/com/netflix/genie/GenieWeb.java
@@ -22,6 +22,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
 import org.springframework.boot.autoconfigure.session.SessionAutoConfiguration;
+import org.springframework.retry.annotation.EnableRetry;
 
 import java.util.Map;
 
@@ -31,6 +32,7 @@ import java.util.Map;
  * @author tgianos
  * @since 3.0.0
  */
+@EnableRetry
 @SpringBootApplication(exclude = {SessionAutoConfiguration.class, RedisAutoConfiguration.class})
 public class GenieWeb {
